move wrap_text to extractor
